% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_surv.R
\name{G_surv}
\alias{G_surv}
\title{Integral of Survival Function of Residual Censoring Variable}
\usage{
G_surv(time, cens, time_Y, weight)
}
\arguments{
\item{time}{The follow up time for right censored data.}

\item{cens}{The censoring status, 0 for censoring, 1 for failure.}

\item{time_Y}{The upper bound of the integral.}

\item{weight}{The weight for each observation.}
}
\value{
This function returns the integral of the residual censoring variable from 0 to a user-specified upper bound.
}
\description{
Integrate the survival function of residual censoring variable from 0 to a user-specified upper bound.
}
\examples{
G_surv(time = rexp(100,1), cens = rbinom(100, 1, 0.8), time_Y = 1.5, weight = rep(1, 100))
G_surv(time = rexp(100,1), cens = rbinom(100, 1, 0.8), time_Y = 1.5, weight = rexp(100, 1))

}
