% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVQRLB_main.R
\name{U_equation}
\alias{U_equation}
\title{Form the Estimating Equation}
\usage{
U_equation(beta, dataset, denom, qtile, weight, norm = TRUE)
}
\arguments{
\item{beta}{The parameter estimate.}

\item{dataset}{The survival data.}

\item{denom}{The denominator used in the estimating equation.}

\item{qtile}{The quantile level used to conduct the quantile regression.}

\item{weight}{The weight of each observation. Default is equal weights.}

\item{norm}{Whether the norm of the function value should be returned. \code{TRUE}: return the norm; \code{FALSE}: return the function value. Default is \code{TRUE}.}
}
\value{
This function returns the function value (a vector of length n+1) of the estimating equation evaluated at "beta" if norm = FALSE; otherwise returns the norm of it.
}
\description{
Set up an estimating equation.
}
\references{
Cai, Z. and Sit, T. (2018+),
"Quantile regression model with time-varying covariates under length-biased sampling,"
\emph{Working Paper}.
}
