% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVQRLB.R, R/TVQRLB_main.R
\docType{package}
\name{TVQRLB}
\alias{TVQRLB}
\alias{TVQRLB-package}
\title{TVQRLB: Quantile Regression Model with Time-Varying Covariates under Length-Biased Sampling}
\usage{
TVQRLB(dataset, betao, bootstrap_time, qtile)
}
\arguments{
\item{dataset}{The survival data.}

\item{betao}{The initial estimate for the parameter.}

\item{bootstrap_time}{The bootstrapping time for multiplier bootstrap.}

\item{qtile}{The quantile level used to conduct the quantile regression.}
}
\value{
This function returns a list containing three elements: 
\itemize{
  \item beta_est, the estimated value of parameter
  \item mean_bootstrap, the mean of bootstrap estimates
  \item sd_bootstrap, the standard error of bootstrap estimates
}
}
\description{
Quantile Regression Model with Time-Varying Covariates under Length-Biased Sampling
}
\examples{
TVQRLB(fixedCP.cen20, c(-1, 0.5, 1.5), 100, 0.5)
TVQRLB(fixedCP.cen40, c(-1, 0.5, 1.5), 100, 0.5)
TVQRLB(fixedCP.cen60, c(-1, 0.5, 1.5), 100, 0.5)
TVQRLB(exponCP.cen20, c(-1, 0.5, 1.5), 100, 0.5)
TVQRLB(exponCP.cen40, c(-1, 0.5, 1.5), 100, 0.5)
TVQRLB(exponCP.cen60, c(-1, 0.5, 1.5), 100, 0.5)

}
\author{
\strong{Maintainer}: Zexi Cai \email{zxcai@link.cuhk.edu.hk}

Authors:
\itemize{
  \item Tony Sit \email{tonysit@sta.cuhk.edu.hk} [contributor]
}

}
