% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVQRLB_main.R
\name{TVQRLB_OP}
\alias{TVQRLB_OP}
\title{Quantile Regression Model with Time-Varying Covariates under Length-Biased Sampling using OP Method for Standard Error Estimation}
\usage{
TVQRLB_OP(dataset, betao, bootstrap_time, qtile, B_size = 10000)
}
\arguments{
\item{dataset}{The survival data.}

\item{betao}{The initial estimate for the parameter.}

\item{bootstrap_time}{The bootstrapping time for estimation of the "meat" matrix V.}

\item{qtile}{The quantile level used to conduct the quantile regression.}

\item{B_size}{The resampling size for the OP method. Default is \code{10000}.}
}
\value{
This function returns a list containing two elements: 
\itemize{
  \item beta_est, the estimated value of parameter
  \item sd_OP, the standard error of parameter estimates
}
}
\description{
Quantile Regression Model with Time-Varying Covariates under Length-Biased Sampling using OP Method for Standard Error Estimation
}
\examples{
TVQRLB_OP(fixedCP.cen20, c(-1, 0.5, 1.5), 1000, 0.5, 10000)
TVQRLB_OP(fixedCP.cen40, c(-1, 0.5, 1.5), 1000, 0.5, 10000)
TVQRLB_OP(fixedCP.cen60, c(-1, 0.5, 1.5), 1000, 0.5, 10000)
TVQRLB_OP(exponCP.cen20, c(-1, 0.5, 1.5), 1000, 0.5, 10000)
TVQRLB_OP(exponCP.cen40, c(-1, 0.5, 1.5), 1000, 0.5, 10000)
TVQRLB_OP(exponCP.cen60, c(-1, 0.5, 1.5), 1000, 0.5, 10000)

}
